# backend/static/swagger.yaml
swagger: "2.0"
info:
  title: Device register API
  version: 1.0.1
  description: API documentation for device register application.
...
paths:
...
  /events/{event_id}:
    get:
      tags:
        - Events
      summary: Get event by ID
      security:
        - bearerAuth: [ ]
      parameters:
        - name: event_id
          in: path
          required: true
          type: integer
      responses:
        '200':
          description: A single event
          schema:
            $ref: '#/definitions/Event'
        '401':
          description: Unauthorized
        '404':
          description: Event not found
    delete:
      tags:
        - Events
      summary: Delete event by ID
      security:
        - bearerAuth: [ ]
      parameters:
        - name: event_id
          in: path
          required: true
          type: integer
      responses:
        '200':
          description: Event deleted
        '401':
          description: Unauthorized
        '404':
          description: Event not found

    patch:
      tags:
        - Events
      summary: Update event by ID
      security:
        - bearerAuth: [ ]
      parameters:
      - name: event_id
        in: path
        required: true
        type: integer
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/NewEvent'
      responses:
        '200':
          description: Event updated
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '404':
          description: Event not found
        '500':
          description: Database error
...
securityDefinitions:
    bearerAuth:
        type: apiKey
        name: Authorization
        in: header
        description: "Enter 'Bearer <JWT>' to authorize. Example: 'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'"
definitions:
...
  NewUser:
    type: object
    properties:
      user_name:
        type: string
      user_email:
        type: string
  Event:
    type: object
    properties:
      event_id:
        type: integer
        example: 1
      dev_id:
        type: integer
        example: 2
      move_time:
        type: string
        format: date-time
        example: "2024-12-11T11:11:11"
      loc_name:
        type: string
        example: "Room 101"
      user_id:
        type: integer
        example: 1
      user_name:
        type: string
        example: "Testi Testinen"
      user_email:
        type: string
        example: "email@email.com"
      company:
        type: string
        example: "Oy Firma Ab"
      comment:
        type: string
        example: "Rikki"
  NewEvent:
    type: object
    properties:
      dev_id:
        type: integer
        example: 1
      move_time:
        type: string
        format: date-time
        example: "2024-12-11 11:11:11"
      loc_name:
        type: string
      company:
        type: string
      comment:
        type: string
      user:
        $ref: '#/definitions/NewUser'
...